Renaming packages and removing unnecessary files
