+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/data/preferences/PreferencesHelper.kt
@@ -1,7 +1,7 @@
 package mustafaozhan.github.com.mycurrencies.data.preferences
 
 import com.google.gson.Gson
-import mustafaozhan.github.com.mycurrencies.base.BaseSharedPreferences
+import mustafaozhan.github.com.mycurrencies.base.BasePreferences
 import mustafaozhan.github.com.mycurrencies.model.Currencies
 import mustafaozhan.github.com.mycurrencies.model.MainData
 import javax.inject.Inject
@@ -11,8 +11,8 @@ import javax.inject.Singleton
  * Created by Mustafa Ozhan on 7/10/18 at 9:50 PM on Arch Linux wit Love <3.
  */
 @Singleton
-class GeneralSharedPreferences @Inject
-constructor() : BaseSharedPreferences() {
+class PreferencesHelper @Inject
+constructor() : BasePreferences() {
 
     companion object {
         const val GENERAL_SHARED_PREFS = "GENERAL_SHARED_PREFS"
