+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/data/repository/BackendRepository.kt
@@ -1,7 +1,7 @@
 package mustafaozhan.github.com.mycurrencies.data.repository
 
 import io.reactivex.Observable
-import mustafaozhan.github.com.mycurrencies.data.api.ApiHelper
+import mustafaozhan.github.com.mycurrencies.data.backend.BackendHelper
 import mustafaozhan.github.com.mycurrencies.model.Currencies
 import mustafaozhan.github.com.mycurrencies.model.CurrencyResponse
 import javax.inject.Inject
@@ -9,7 +9,7 @@ import javax.inject.Singleton
 
 @Singleton
 class BackendRepository @Inject
-constructor(private val apiHelper: ApiHelper) {
+constructor(private val backendHelper: BackendHelper) {
     fun getAllOnBase(base: Currencies): Observable<CurrencyResponse> =
-        apiHelper.apiService.getAllOnBase(base)
+        backendHelper.backendService.getAllOnBase(base)
 }
