+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/data/repository/PreferencesRepository.kt
@@ -1,20 +1,20 @@
 package mustafaozhan.github.com.mycurrencies.data.repository
 
-import mustafaozhan.github.com.mycurrencies.data.preferences.GeneralSharedPreferences
+import mustafaozhan.github.com.mycurrencies.data.preferences.PreferencesHelper
 import mustafaozhan.github.com.mycurrencies.model.MainData
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class PreferencesRepository @Inject
-constructor(private val generalSharedPreferences: GeneralSharedPreferences) {
+constructor(private val sharedPreferencesHelper: PreferencesHelper) {
 
-    fun loadMainData() = generalSharedPreferences.loadMainData()
+    fun loadMainData() = sharedPreferencesHelper.loadMainData()
 
     fun persistMainData(mainData: MainData) =
-        generalSharedPreferences.persistMainData(mainData)
+        sharedPreferencesHelper.persistMainData(mainData)
 
-    fun loadResetData() = generalSharedPreferences.loadResetData()
+    fun loadResetData() = sharedPreferencesHelper.loadResetData()
 
-    fun persistResetData(resetData: Boolean) = generalSharedPreferences.persistResetData(resetData)
+    fun persistResetData(resetData: Boolean) = sharedPreferencesHelper.persistResetData(resetData)
 }
