+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/di/ApplicationModule.kt
@@ -3,10 +3,6 @@ package mustafaozhan.github.com.mycurrencies.di
 import dagger.Module
 import dagger.Provides
 import mustafaozhan.github.com.mycurrencies.room.AppDatabase
-import mustafaozhan.github.com.mycurrencies.room.AppExecutors
-import mustafaozhan.github.com.mycurrencies.room.DiskIOThreadExecutor
-import mustafaozhan.github.com.mycurrencies.room.MainThreadExecutor
-import java.util.concurrent.Executors
 import javax.inject.Singleton
 
 @Module
@@ -24,13 +20,4 @@ class ApplicationModule {
     @Singleton
     internal fun providesOfflineRatesDao(database: AppDatabase) =
         database.offlineRatesDao()
-
-    @Singleton
-    @Provides
-    internal fun providesAppExecutors() =
-        AppExecutors(
-            DiskIOThreadExecutor(),
-            Executors.newFixedThreadPool(AppExecutors.THREAD_COUNT),
-            MainThreadExecutor()
-        )
 }
