+++ app/src/main/res/layout/layout_main_toolbar.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
+    android:id="@+id/app_bar_layout"
     android:layout_height="wrap_content"
     android:background="@color/colorPrimaryDark">
