+++ app/src/main/res/layout/layout_settings_toolbar.xml
@@ -2,6 +2,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
+    android:id="@+id/app_bar_layout"
     android:layout_height="wrap_content"
     android:background="@color/colorPrimaryDark">
